Move lfs_stag_t to lfs.h in preparation for adjusting types for 16 bit cpus
Some functions return either a negative error code or the tag for a file
or directory. On 32 bit machines int and lfs_stag_t are the same, so we
can just use int in the .h file, but for 16 bit machines int is 16 bits
and we end up with fatal errors due to type conversion errors.