lfs_remove: Use lfs_stag_t as return code for lfs_remove
lfs_remove returns either a negative return code or a tag.
For 32 bit machines with int as 32 bits this co-incides, but for smaller
bit processors, we need to ensure a 32 bit value is returned so change
the return type to lfs_stag_t and cast any error codes to it.