lfs_file_rewind: Use lfs_soff_t as return code for lfs_file_rewind
lfs_file_rewind returns either a negative return code or an offset.
For 32 bit machines with int as 32 bits this co-incides, but for smaller
bit processors, we need to ensure a 32 bit value is returned so change
the return type to lfs_soff_t and cast any error codes to it.