lfs_file_sync: Use lfs_ssize_t as return code for lfs_file_sync
lfs_file_sync returns either a negative return code or a size.
For 32 bit machines with int as 32 bits this co-incides, but for smaller
bit processors, we need to ensure a 32 bit value is returned so change
the return type to lfs_ssize_t and cast any error codes to it.