Adjust lfs_dir_find return code to ensure 32 bit value.
lfs_dir_find returns either a negative return code or a size.
For 32 bit machines with int as 32 bits this co-incides, but for smaller
bit processors, we need to ensure a 32 bit value is returned so change
the return type to lfs_stag_t and cast any error codes to it.