Support no validation to improve write performance
It can improve write performance if no validation, but drivers
must ensure data integrity.

I test lfs on spinand with NFTL. The spinand has 128M capacity
with 2K page size and 128K block size. The actually page size
changes to 4K on my using NFTL. And lfs is configured as follows:

    .read_size = 4096
    .write_szie = 4096
    .block_szie = 4096
    .block_count = 25680
    .block_cycles = -1
    .lookahead_size = 256

The performance on raw device without littlefs:

    read average speed: 14361.85 KB/s
    write average speed: 2994.74 KB/s

The performance on lfs with validation as follow:

    read average speed: 12094.49 KB/s
    write average speed: 2074.97 KB/s

The performance on lfs without validation as follow:

    read average speed: 12427.18 KB/s
    write average speed: 2674.33 KB/s

It really improves 600 KB/s to write if disable validation.
To me, the NFTL has ensure data integrity, there is no need
to validate again on lfs. It is also helpfull to flash like
MMC.